src/api/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bed-management-api"
version = "0.1.0"
description = "Patient Flow / Bed Management — Agentic AI Demo API"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]",
    "uvicorn[standard]",
    "azure-ai-projects>=2.0.0b1",
    "azure-identity",
    "pydantic>=2.0",
    "pydantic-settings",
    "sse-starlette",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
    "ruff",
]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "httpx>=0.27",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 120
